requirements.txt
# Core ML/DL frameworks
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20

# Vector stores and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Data processing
datasets>=2.15.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
sentencepiece>=0.1.99
tiktoken>=0.5.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8

# Logging and monitoring
wandb>=0.16.0  # Optional, for experiment tracking
tensorboard>=2.15.0

# API (for deployment)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0
